use interceptor instead of eventing
