+++ packages/examples/src/relay/Feedback/Feedback.stories.tsx
@@ -5,7 +5,7 @@ import {
   MockPayloadGenerator as PayloadGenerator,
   type StoryObjWithoutFragmentRefs,
   type WithNovaEnvironment,
-  EventingProvider,
+  EventingInterceptor,
 } from "@nova/react-test-utils/relay";
 import type { Meta } from "@storybook/react";
 import { userEvent, waitFor, within, expect } from "@storybook/test";
@@ -135,12 +135,16 @@ const FeedbackWithDeleteDialog = (
   const [open, setOpen] = React.useState(false);
   const [text, setText] = React.useState("");
   return (
-    <EventingProvider<typeof events>
+    <EventingInterceptor<typeof events>
       eventMap={{
         onDeleteFeedback: (eventWrapper) => {
           setOpen(true);
           setText(eventWrapper.event.data().feedbackText);
-          return Promise.resolve();
+          return Promise.resolve(undefined);
+        },
+        feedbackTelemetry: (eventWrapper) => {
+          console.log("Telemetry event", eventWrapper.event.data());
+          return Promise.resolve(eventWrapper);
         },
       }}
     >
@@ -149,7 +153,7 @@ const FeedbackWithDeleteDialog = (
         <button onClick={() => setOpen(false)}>Cancel</button>
         Are you sure you want to delete feedback "{text}"
       </dialog>
-    </EventingProvider>
+    </EventingInterceptor>
   );
 };
