+++ packages/examples/src/relay/pure-relay/Feedback.stories.tsx
@@ -3,7 +3,7 @@ import {
   getNovaDecorator,
   getNovaEnvironmentForStory,
   type WithNovaEnvironment,
-  EventingProvider,
+  EventingInterceptor,
   getOperationName,
   getOperationType,
   type StoryObjWithoutFragmentRefs,
@@ -203,12 +203,16 @@ const FeedbackWithDeleteDialog = (
   const [open, setOpen] = React.useState(false);
   const [text, setText] = React.useState("");
   return (
-    <EventingProvider<typeof events>
+    <EventingInterceptor<typeof events>
       eventMap={{
         onDeleteFeedback: (eventWrapper) => {
           setOpen(true);
           setText(eventWrapper.event.data().feedbackText);
-          return Promise.resolve();
+          return Promise.resolve(undefined);
+        },
+        feedbackTelemetry: (eventWrapper) => {
+          console.log("Telemetry event", eventWrapper.event.data());
+          return Promise.resolve(eventWrapper);
         },
       }}
     >
@@ -217,7 +221,7 @@ const FeedbackWithDeleteDialog = (
         <button onClick={() => setOpen(false)}>Cancel</button>
         Are you sure you want to delete feedback "{text}"
       </dialog>
-    </EventingProvider>
+    </EventingInterceptor>
   );
 };
